SFT-154 SFR-160 SFR-159 SFR-161 Initial Commit
This includes basic functionality for an ingest process that brings
metadata and .epub files from Project Gutenberg into a local postgres
database and an AWS ElasticSearch instance. The data is relatively
rough at the moment, but it fits the bibframe model and provides a test
base on which the project data can be examined and further tests run.

The data is (hypothetically) enhanced from the MetadataWrangler service
as well as the OCLC APIs, though this does not always bring in complete data.

This model can also be improved by assigning further identifiers to the
records generated and handling the full complexity of the resulting records
in a better way. These tasks are documented in the code and will be written
up.